fix: change file structure
