+++ src/components/Main/Countries_list/countries_list.scss
