+++ src/components/Main/Map/Graph/graph.scss
