+++ src/components/Main/Map/Map.scss
@@ -1,0 +1,9 @@
+@import '../../../../src/shared/variables';
+
+.map {
+  width: 60%;
+  margin: 5px;
+  border: 1px solid $gray-700;
+  background-color: $gray-800;
+  border-radius: 5px;
+}
