--- src/components/Main/Statistic/statistic.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import GlobalStatistic from '../GlobalStatistic/global-statistic';
-import TestsStatistic from '../TestsStatistic/tests-statistic';
-import Graph from '../Graph/graph';
-import './statistic.scss';
-
-const Statistic = () => (
-  <section className="statistic">
-    <GlobalStatistic />
-    <TestsStatistic />
-    <Graph />
-  </section>
-);
-
-export default Statistic;
