+++ src/components/Main/Table/table.scss
