--- src/components/Map/Map.scss
@@ -1,9 +1,0 @@
-@import '../../shared/variables';
-
-.map {
-  width: 60%;
-  margin: 5px;
-  border: 1px solid $gray-700;
-  background-color: $gray-800;
-  border-radius: 5px;
-}
